optimize parameters and logic
