+++ pom.xml
@@ -35,7 +35,7 @@
         <module>kyuubi-main</module>
         <module>kyuubi-metrics</module>
         <module>kyuubi-zookeeper</module>
-        <module>tools/kubernetes/cache-file-cleaner</module>
+        <module>tools/kubernetes/shuffle-file-cleaner</module>
     </modules>
     <packaging>pom</packaging>
 
@@ -66,7 +66,6 @@
         <codahale.metrics.version>4.1.1</codahale.metrics.version>
         <commons-codec.version>1.15</commons-codec.version>
         <commons-collections.version>3.2.2</commons-collections.version>
-        <commons-cli.version>1.3.1</commons-cli.version>
         <commons-lang3.version>3.10</commons-lang3.version>
         <curator.version>2.12.0</curator.version>
         <delta.version>0.8.0</delta.version>
@@ -287,12 +286,6 @@
                 <groupId>commons-collections</groupId>
                 <artifactId>commons-collections</artifactId>
                 <version>${commons-collections.version}</version>
-            </dependency>
-
-            <dependency>
-                <groupId>commons-cli</groupId>
-                <artifactId>commons-cli</artifactId>
-                <version>${commons-cli.version}</version>
             </dependency>
 
             <dependency>
