+++ tools/kubernetes/docker/shuffle-file-cleaner/Dockerfile
