+++ tools/kubernetes/docker/shuffle-file-cleaner/entrypoint.sh
@@ -15,24 +15,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-var CMD = "java -cp ${CLEANER_CLASSPATH} org.apache.kyuubi.KubernetesShuffleFileCleaner  "
-
-CMD="${CMD} -cache_dirs ${SHUFFLE_DIR}"
-
-if ! [[ -n "${FILE_EXPIRED_TIME}" ]]; then
-  CMD="${CMD} -file_expired_time ${FILE_EXPIRED_TIME}"
-fi
-
-if ! [[ -n "${FREE_SPACE_THRESHOLD}" ]]; then
-  CMD="${CMD} -free_space_threshold ${FREE_SPACE_THRESHOLD}"
-fi
-
-if ! [[ -n "${CYCLE_INTERVAL_TIME}" ]]; then
-  CMD="${CMD} -cycle_interval_time ${CYCLE_INTERVAL_TIME}"
-fi
-
-if ! [[ -n "${DEEP_CLEAN_FILE_EXPIRED_TIME}" ]]; then
-  CMD="${CMD} -deep_clean_file_expired_time ${DEEP_CLEAN_FILE_EXPIRED_TIME}"
-fi
-
-${CMD}
+java -cp ${CLEANER_CLASSPATH} org.apache.kyuubi.KubernetesShuffleFileCleaner
