+++ tools/kubernetes/shuffle-file-cleaner/pom.xml
@@ -26,8 +26,8 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>cache-file-cleaner</artifactId>
-    <name>kyuubi tool for clean cache files on kubernetes</name>
+    <artifactId>shuffle-file-cleaner</artifactId>
+    <name>Kyuubi tool Kubernetes shuffle cleaner</name>
     <packaging>jar</packaging>
 
     <dependencies>
@@ -35,11 +35,6 @@
         <dependency>
             <groupId>commons-cli</groupId>
             <artifactId>commons-cli</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-log4j12</artifactId>
         </dependency>
 
     </dependencies>
