+++ tools/kubernetes/shuffle-file-cleaner/src/main/resources/log4j-defaults.properties
